Bank Transaction Fraud Analytics/requirements.txt
streamlit
snowflake-connector-python[pandas]
pandas
plotly
python-dotenv